feat: added a single player page and fixed the animated backgroud issue in home page
