+++ src/App.tsx
@@ -1,6 +1,7 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom"
 import Layout from "./pages/Layout"
 import Home from "./pages/Home"
+import SinglePlayer from "./pages/SinglePlayer"
 import './App.css'
 
 function App() {
@@ -10,6 +11,7 @@ function App() {
       <Routes>
         <Route path="/" element={<Layout />}>
           <Route index element={<Home />} />
+          <Route path="SinglePlayer" element={<SinglePlayer />} />
         </Route>
       </Routes>
     </BrowserRouter>
